+++ components/issues/controllers/GenerateResolutionPlanController.tsx
@@ -2,7 +2,7 @@
 
 import { toast } from "@/hooks/use-toast"
 import { CommentRequestSchema } from "@/lib/schemas/api"
-import { getApiKeyFromLocalStorage, SSEUtils } from "@/lib/utils"
+import { getApiKeyFromLocalStorage, SSEUtils } from "@/lib/utils/utils-common"
 
 interface Props {
   issueNumber: number
