+++ components/ui/accordion.tsx
@@ -4,7 +4,7 @@ import * as AccordionPrimitive from "@radix-ui/react-accordion"
 import { ChevronDown } from "lucide-react"
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/utils-common"
 
 const Accordion = AccordionPrimitive.Root
