+++ components/ui/navigation-menu.tsx
@@ -3,7 +3,7 @@ import { cva } from "class-variance-authority"
 import { ChevronDown } from "lucide-react"
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/utils-common"
 
 const NavigationMenu = React.forwardRef<
   React.ElementRef<typeof NavigationMenuPrimitive.Root>,
