+++ components/ui/pagination.tsx
@@ -3,7 +3,7 @@ import Link from "next/link"
 import * as React from "react"
 
 import { ButtonProps, buttonVariants } from "@/components/ui/button"
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/utils-common"
 
 const Pagination = ({ className, ...props }: React.ComponentProps<"nav">) => (
   <nav
