+++ components/ui/tabs.tsx
@@ -3,7 +3,7 @@
 import * as TabsPrimitive from "@radix-ui/react-tabs"
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/utils-common"
 
 const Tabs = TabsPrimitive.Root
