+++ lib/utils/utils-common.ts
