+++ lib/workflows/commentOnIssue.ts
@@ -11,7 +11,7 @@ import { updateJobStatus } from "@/lib/redis-old"
 import { SearchCodeTool } from "@/lib/tools"
 import GetFileContentTool from "@/lib/tools/GetFileContent"
 import { GitHubRepository } from "@/lib/types/github"
-import { setupLocalRepository } from "@/lib/utils-server"
+import { setupLocalRepository } from "@/lib/utils/utils-server"
 
 interface GitHubError extends Error {
   status?: number
